moved file to components and made cursor turn into pointer on hover
